requirements.txt
streamlit
requests
pillow
pyyaml